gain and offset are doubles

+++ GIP/gip/GeoImage.h
@@ -198,9 +198,9 @@ namespace gip {
 
         //! \name Multiple band convenience functions
         //! Set gain for all bands
-        void SetGain(float gain) { for (unsigned int i=0;i<_RasterBands.size();i++) _RasterBands[i].SetGain(gain); }
+        void SetGain(double gain) { for (unsigned int i=0;i<_RasterBands.size();i++) _RasterBands[i].SetGain(gain); }
         //! Set gain for all bands
-        void SetOffset(float offset) { for (unsigned int i=0;i<_RasterBands.size();i++) _RasterBands[i].SetOffset(offset); }
+        void SetOffset(double offset) { for (unsigned int i=0;i<_RasterBands.size();i++) _RasterBands[i].SetOffset(offset); }
         //! Set  for all bands
         void SetUnits(std::string units) { for (unsigned int i=0;i<_RasterBands.size();i++) _RasterBands[i].SetUnits(units); }
         //! Set NoData for all bands
